requirements.txt
openai
python-dotenv
google-generativeai
rich
pytest
